+++ client/app/visualizations/choropleth/ColorPalette.js
@@ -1,0 +1,8 @@
+import { extend } from 'lodash';
+import ColorPalette from '@/visualizations/ColorPalette';
+
+export default extend({
+  White: '#ffffff',
+  Black: '#000000',
+  'Light Gray': '#dddddd',
+}, ColorPalette);
