+++ client/app/visualizations/choropleth/Renderer/index.jsx
@@ -1,14 +1,15 @@
-import { isEqual, omit, merge } from 'lodash';
-import React, { useState, useEffect, useRef, useMemo } from 'react';
+import { omit, merge } from 'lodash';
+import React, { useState, useEffect } from 'react';
 import { RendererPropTypes } from '@/visualizations';
 import { $http } from '@/services/ng';
+import useMemoWithDeepCompare from '@/lib/hooks/useMemoWithDeepCompare';
 
 import initChoropleth from './initChoropleth';
 import { prepareData } from './utils';
 import './renderer.less';
 
-import countriesDataUrl from './maps/countries.geo.json';
-import subdivJapanDataUrl from './maps/japan.prefectures.geo.json';
+import countriesDataUrl from '../maps/countries.geo.json';
+import subdivJapanDataUrl from '../maps/japan.prefectures.geo.json';
 
 function getDataUrl(type) {
   switch (type) {
@@ -18,53 +19,9 @@ function getDataUrl(type) {
   }
 }
 
-const geoJsonCache = {};
-
-function useGeoJson(url) {
-  const [data, setData] = useState(null);
-
-  useEffect(() => {
-    if (!geoJsonCache[url]) {
-      geoJsonCache[url] = {
-        url,
-        promise: $http.get(url).then(response => response.data),
-        refcount: 0,
-      };
-    }
-    const item = geoJsonCache[url];
-    item.refcount += 1;
-    let cancelled = false;
-    item.promise.then((result) => {
-      if (!cancelled) {
-        setData(result);
-      }
-    });
-
-    return () => {
-      cancelled = true;
-      item.refcount -= 1;
-      if (item.refcount <= 0) {
-        geoJsonCache[item.url] = undefined;
-      }
-    };
-  }, [url]);
-
-  return data;
-}
-
-function useMemoWithDeepCompare(create, inputs) {
-  const valueRef = useRef();
-  const value = useMemo(create, inputs);
-  if (!isEqual(value, valueRef.current)) {
-    valueRef.current = value;
-  }
-  return valueRef.current;
-}
-
 export default function Renderer({ data, options, onOptionsChange }) {
   const [container, setContainer] = useState(null);
-
-  const geoJson = useGeoJson(getDataUrl(options.mapType));
+  const [geoJson, setGeoJson] = useState(null);
 
   const optionsWithoutBounds = useMemoWithDeepCompare(
     () => omit(options, ['bounds']),
@@ -72,6 +29,18 @@ export default function Renderer({ data, options, onOptionsChange }) {
   );
 
   const [map, setMap] = useState(null);
+
+  useEffect(() => {
+    let cancelled = false;
+
+    $http.get(getDataUrl(options.mapType)).then((response) => {
+      if (!cancelled) {
+        setGeoJson(response.data);
+      }
+    });
+
+    return () => { cancelled = true; };
+  }, [options.mapType]);
 
   useEffect(() => {
     if (container) {
