+++ client/app/visualizations/choropleth/Renderer/renderer.less
@@ -1,5 +1,6 @@
 .choropleth-visualization-legend {
-  padding: 5px;
+  padding: 3px;
+  cursor: default;
 
   > div {
     line-height: 1;
