+++ client/app/visualizations/choropleth/Renderer/utils.js
@@ -1,13 +1,6 @@
+import { isString, isObject, isFinite, each, map, extend, uniq, filter, first } from 'lodash';
 import chroma from 'chroma-js';
-import _ from 'lodash';
 import { createNumberFormatter as createFormatter } from '@/lib/value-format';
-import DefaultPalette from '@/visualizations/ColorPalette';
-
-export const ColorPalette = _.extend({
-  White: '#ffffff',
-  Black: '#000000',
-  'Light Gray': '#dddddd',
-}, DefaultPalette);
 
 export function darkenColor(color) {
   return chroma(color).darken().hex();
@@ -16,7 +9,7 @@ export function darkenColor(color) {
 export function createNumberFormatter(format, placeholder) {
   const formatter = createFormatter(format);
   return (value) => {
-    if (_.isNumber(value) && isFinite(value)) {
+    if (isFinite(value)) {
       return formatter(value);
     }
     return placeholder;
@@ -29,7 +22,7 @@ export function prepareData(data, countryCodeField, valueField) {
   }
 
   const result = {};
-  _.each(data, (item) => {
+  each(data, (item) => {
     if (item[countryCodeField]) {
       const value = parseFloat(item[valueField]);
       result[item[countryCodeField]] = {
@@ -44,24 +37,24 @@ export function prepareData(data, countryCodeField, valueField) {
 
 export function prepareFeatureProperties(feature, valueFormatted, data, countryCodeType) {
   const result = {};
-  _.each(feature.properties, (value, key) => {
+  each(feature.properties, (value, key) => {
     result['@@' + key] = value;
   });
   result['@@value'] = valueFormatted;
   const datum = data[feature.properties[countryCodeType]] || {};
-  return _.extend(result, datum.item);
+  return extend(result, datum.item);
 }
 
 export function getValueForFeature(feature, data, countryCodeType) {
   const code = feature.properties[countryCodeType];
-  if (_.isString(code) && _.isObject(data[code])) {
+  if (isString(code) && isObject(data[code])) {
     return data[code].value;
   }
   return undefined;
 }
 
 export function getColorByValue(value, limits, colors, defaultColor) {
-  if (_.isNumber(value) && isFinite(value)) {
+  if (isFinite(value)) {
     for (let i = 0; i < limits.length; i += 1) {
       if (value <= limits[i]) {
         return colors[i];
@@ -73,9 +66,9 @@ export function getColorByValue(value, limits, colors, defaultColor) {
 
 export function createScale(features, data, options) {
   // Calculate limits
-  const values = _.uniq(_.filter(
-    _.map(features, feature => getValueForFeature(feature, data, options.countryCodeType)),
-    _.isNumber,
+  const values = uniq(filter(
+    map(features, feature => getValueForFeature(feature, data, options.countryCodeType)),
+    isFinite,
   ));
   if (values.length === 0) {
     return {
@@ -91,7 +84,7 @@ export function createScale(features, data, options) {
       colors: [options.colors.max],
       legend: [{
         color: options.colors.max,
-        limit: _.first(values),
+        limit: first(values),
       }],
     };
   }
@@ -102,45 +95,10 @@ export function createScale(features, data, options) {
     .colors(limits.length);
 
   // Group values for legend
-  const legend = _.map(colors, (color, index) => ({
+  const legend = map(colors, (color, index) => ({
     color,
     limit: limits[index],
   })).reverse();
 
   return { limits, colors, legend };
-}
-
-export function inferCountryCodeType(mapType, data, countryCodeField) {
-  const regexMap = {
-    countries: {
-      iso_a2: /^[a-z]{2}$/i,
-      iso_a3: /^[a-z]{3}$/i,
-      iso_n3: /^[0-9]{3}$/i,
-    },
-    subdiv_japan: {
-      name: /^[a-z]+$/i,
-      name_local: /^[\u3400-\u9FFF\uF900-\uFAFF]|[\uD840-\uD87F][\uDC00-\uDFFF]+$/i,
-      iso_3166_2: /^JP-[0-9]{2}$/i,
-    },
-  };
-
-  const regex = regexMap[mapType];
-
-  const initState = _.mapValues(regex, () => 0);
-
-  const result = _.chain(data)
-    .reduce((memo, item) => {
-      const value = item[countryCodeField];
-      if (_.isString(value)) {
-        _.each(regex, (r, k) => {
-          memo[k] += r.test(value) ? 1 : 0;
-        });
-      }
-      return memo;
-    }, initState)
-    .toPairs()
-    .reduce((memo, item) => (item[1] > memo[1] ? item : memo))
-    .value();
-
-  return (result[1] / data.length) >= 0.9 ? result[0] : null;
 }
